requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
chardet==2.3.0
Django==2.2.1
docx2txt==0.6
EbookLib==0.15
lxml==4.3.3
pdfminer.six==20181108
Pillow==6.0.0
pycryptodome==3.8.1
PyPDF2==1.26.0
pytesseract==0.2.6
python-pptx==0.6.5
pytz==2019.1
six==1.10.0
sortedcontainers==2.1.0
SpeechRecognition==3.6.3
sqlparse==0.3.0
xlrd==1.0.0
XlsxWriter==1.1.8
